@@ -18,17 +18,19 @@
         "gitlab"
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">=8.0",
         "league/oauth2-client": "^2.4.1"
     },
     "require-dev": {
-        "friendsofphp/php-cs-fixer": "^2.0",
-        "guzzlehttp/psr7": "^1.6",
-        "http-interop/http-factory-guzzle": "^1.0",
-        "mockery/mockery": "^1.0",
-        "m4tthumphrey/php-gitlab-api": "^10.0|^11.0",
-        "php-http/guzzle7-adapter": "^0.1",
-        "phpunit/phpunit": "^8.0|^9.0"
+        "guzzlehttp/psr7": "^2.4.1",
+        "http-interop/http-factory-guzzle": "^1.2",
+        "mockery/mockery": "^1.5.1",
+        "m4tthumphrey/php-gitlab-api": "^10.0|^11.8",
+        "php-http/guzzle7-adapter": "^1.0.0",
+        "phpunit/phpunit": "^8.0|^9.5.25",
+        "phpstan/phpstan": "^1.8.10",
+        "phpstan/phpstan-mockery": "^1.1",
+        "phpstan/extension-installer": "^1.2"
     },
     "suggest": {
         "m4tthumphrey/php-gitlab-api": "For further API usage using the acquired OAuth2 token"
@@ -46,6 +48,11 @@
     "extra": {
         "branch-alias": {
             "dev-master": "3.x-dev"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "phpstan/extension-installer": true
         }
     }
 }
